+++ apps/meteor/ee/server/lib/omnichannel/routing/LoadRotation.ts
@@ -1,7 +1,7 @@
 import type { IOmnichannelCustomAgent } from '@rocket.chat/core-typings';
 import { Users } from '@rocket.chat/models';
 
-import type { IRoutingManagerConfig } from '../../../../../definition/IRoutingManagerConfig';
+import type { IRoutingManagerConfig } from './IRoutingManagerConfig';
 import { RoutingManager } from '../../../../../server/lib/omnichannel/RoutingManager';
 import { settings } from '../../../../../server/settings';
 import { getChatLimitsQuery } from '../../../hooks/omnichannel/applySimultaneousChatsRestrictions';
